requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
console==0.9907
ezenv==0.92
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
langdetect==1.0.9
lib==4.0.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
six==1.16.0
sympy==1.12
tokenizers==0.14.1
torch==2.0.1
tqdm==4.66.1
transformers==4.35.1
typing_extensions==4.8.0
urllib3==2.1.0